+++ Assets/Art/Battle/Enemies/FireGoon.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e832bcf4e95b1204c8ead1122ba4399f
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
