--- Assets/Art/Battle/Enemies/TornadoGoon/Canvas.controller.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 83b518c152c055949b6686b6f63d15e1
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 9100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
