--- Assets/Art/Battle/Enemies/TornadoGoon/None.anim.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: dfe6ae3a4047e444aac6a17daffb4d8a
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 7400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
