--- Assets/Art/Battle/Enemies/TornadoGoon/Swing.anim.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: acf2d6674bd7a574a9e081a4f6c31476
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 7400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
