--- Assets/EnterFactory.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 85193416f27e62e488bf268b180d200f
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
