+++ Assets/Prefab/Combat/Enemies/MagmaGoon.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: e0d16c1217048f74ab2a3e72f1c32570
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
