+++ Assets/Scripts/Utility/Doorway.cs
@@ -1,0 +1,17 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+using UnityEngine.SceneManagement;
+
+public class Doorway : MonoBehaviour
+{
+    public string Scene;
+
+    private void OnTriggerEnter2D(Collider2D collision)
+    {
+        if (collision.tag == "Player")
+        {
+            SceneManager.LoadScene(Scene);
+        }
+    }
+}
