+++ Assets/Scripts/Utility/Doorway.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: f71843b8e98fd0b439ab3c23d8e891a9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
