+++ Assets/Scripts/Utility/WallAdder.cs
@@ -1,0 +1,17 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class WallAdder : MonoBehaviour
+{
+    public GameObject Wall;
+    
+    private void OnTriggerEnter2D(Collider2D collision)
+    {
+        if (collision.tag == "Player")
+        {
+            Wall.SetActive(true);
+            Destroy(gameObject);
+        }
+    }
+}
