--- Assets/Scripts/WallRemover.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: b026510871363e24581c4a6ddc92d171
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
