--- Assets/WallAdder.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 1bf3f158f56739c43abd20fb656f1815
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
