+++ ProjectSettings/EditorBuildSettings.asset
@@ -17,6 +17,12 @@ EditorBuildSettings:
   - enabled: 1
     path: Assets/Scenes/AirFactory.unity
     guid: 3504179acac10b6408c4da80aa87aadb
+  - enabled: 1
+    path: Assets/Scenes/FireIslandTileMap.unity
+    guid: b158a7e1a4289814cbbf0591abb6fe7b
+  - enabled: 1
+    path: Assets/Scenes/FireColosseum.unity
+    guid: d15f77abfe7083d41bf358c1464b8bec
   m_configObjects:
     com.unity.adaptiveperformance.loader_settings: {fileID: 11400000, guid: 1d42c9c49388c1640b16a8bbfc983157, type: 2}
     com.unity.adaptiveperformance.simulator.provider_settings: {fileID: 11400000, guid: 3eb05e13d47e94045a9f067f9f79e40c, type: 2}
