+++ UserSettings/EditorUserSettings.asset
@@ -21,13 +21,13 @@ EditorUserSettings:
       value: 0109010754045e5e545d092312710a44424e4879747870627f2f4d63b5e13639
       flags: 0
     RecentlyUsedSceneGuid-5:
-      value: 07000100520208590b0b5b7248705b4447151e7979702561797f4931bae2313b
+      value: 500404525402505a0e0f0f7340210944464f1b7e28297e602c284564e3e1303a
       flags: 0
     RecentlyUsedSceneGuid-6:
       value: 0100015e04020c0a0c5a5e7a497b0e4415151a2c7c7d7f612c2b1f65e4e5633a
       flags: 0
     RecentlyUsedSceneGuid-7:
-      value: 500404525402505a0e0f0f7340210944464f1b7e28297e602c284564e3e1303a
+      value: 07000100520208590b0b5b7248705b4447151e7979702561797f4931bae2313b
       flags: 0
     RecentlyUsedSceneGuid-8:
       value: 55535055515659595f575c2749200744474e4f2e7a782765747a4b62e6b9616c
